pyproject.toml
[project]
name = "job-seeker"
version = "0.1.0"
description = "Job Scraper - Search remote jobs across multiple ATS platforms"
authors = [
    {name = "Davi Oliveira", email = "davioliveira353.do@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "weasyprint (>=61.0,<62.0)",
    "pydyf (>=0.10.0,<0.11.0)",
    "flask (>=3.1.0,<4.0.0)"
]

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry]
packages = [{include = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
